feat(windows): port refine, sync, insights and the HTTP client to Spit.Core
SkipGate, Budget, Outbox, RefineService (IRefiner), SyncService with the rule 46 hotkey echo, HttpVoiceApiClient, HeatmapGrid, InsightsPresentation/InsightsFormat, InsightsCache, InsightsModel and IanaTimeZone (rule 43). Ports SkipGateTests (16), BudgetTests (1), OutboxTests (1), RefineServiceTests (8), SyncServiceTests (3) and InsightsTests (24) with the Mac's test names, plus SyncServiceHotkeyEchoTests, IanaTimeZoneTests and HttpVoiceApiClientTests. InsightsTotals.AudioMs widened to long to match Swift's 64-bit Int.

+++ windows/Spit.Core/Insights/Insights.cs
@@ -5,7 +5,9 @@ namespace Spit.Core;
 // Numbers only, by design — the server never selects `raw` or `cleaned` for this endpoint, so no
 // transcript text can reach this type or the cache file it is written to (rule 25).
 
-public sealed record InsightsTotals(int Dictations, int Words, int AudioMs, int? Wpm);
+/// `AudioMs` is 64-bit like Swift's `Int`: an all-history total passes `int.MaxValue` at ~596 hours of
+/// audio, and an `int` would turn every refresh after that into a decoding error.
+public sealed record InsightsTotals(int Dictations, int Words, long AudioMs, int? Wpm);
 
 public sealed record InsightsStreak(int Current, int Longest);
